Added view and template to 'Country Demo' app
This commit adds templates to the site so there is a default page.
This page has a link to the countrydemo index view.

this shows how you can use the 'url' template tag to link to the app

See more here:
https://docs.djangoproject.com/en/1.11/ref/templates/builtins/#url

in `settings.py` It also adds the Country Demo app to the INSTALLED_APPS and the site
templates directory to  `TEMPLATES`